Parse sharding rule factor sizes as int64_t

+++ shardy/dialect/sdy/ir/parsers.cc
@@ -353,7 +353,7 @@ ParseResult parseFactorSizes(AsmParser& parser,
       return failure();
     }
 
-    int factorSize;
+    int64_t factorSize;
     if (parser.parseInteger(factorSize)) {
       return failure();
     }
